Test for PlayerSatOut event only being sent to the table after a hand has ended.
Asumption is that sitting out will eventually put you into post and fold mode, where you can still win if everyone folds before you're to act, or it's checked down and you have the best hand. Currently the second part of this is not yet implemented but may be someday.