Adding support for players sitting out who are not next to act.
Player is marked as sitting out, but will fold if prompted to act. They can still win the hand if everyone else were to fold before their turn. As mentioned previously, functionality to allow them to check the hand down if possible is not yet implemented. (and probably unecessary at this stage)