Fix the payout of the pots. Filter winner by folded first instead of after.
This was auto-formatted to the wrong place, it should have been in the loop so
that it will be run on each of the pots. Also the filter list wasn't being
done correctly so it would attempt to payout a winner that had better cards
but had already folded. This shows up as random exceptions in the existing
tests and will soon have it's own test.