Hookup sit button to actual network seat request/response.
Sit button now claims your seat on the server and disables all sit buttons. Leaving a table, or disconnecting from the server, is not yet handled on either side of the connection so things get a little wonky with repeated tests. Best to kill off server and client each time at the moment.

Disconnections and leaving tables coming up next.