Add a base client class, and more errbacks.
The base class should have everything you need to implement to be a
client. Plus it'll have defaults for anything you don't handle.

The extra errbacks will allow you to catch errors that could occur
in the success callbacks of the ui.client. So if you mess up you
don't get an ugly stacktrace (well, if you override the log_error).