Hookup GTK client handling for call/raise/fold events.
Change the point in time when PlayerPostedBlind goes out. While in the engine, the table find players willing to post blinds before spawning a game, and the game actually collects those blinds, it makes more sense for the client to see the event when they agree to post rather than when the actual blind is collected. (due to the PlayerPrompted going out to indicate that the table is at least trying to start a new hand)