Begin implementation of ranking hands in the same class
The value of each hand is represented as a 6 digit hex number. The most
significant digit represents the hand class, i.e. 0 for high card, 1, for one
pair. The remaining digits are used to represent hand-class specific details,
but in general the next most significant digit represents the value of the
highest card used to classify the hand (for instance, the face value of the
pair card for one pair), and so forth until we are selecting the highest
remaining singles to make up a hand of five cards.

For example, with pocket cards jc, 8h, and table cards
5d, 4s, jh, 3c, 7h the value is:

0x1A8750
  │││││└ Unused
  ││││└─ 5d
  │││└── 7h
  ││└─── 8h
  │└──── jh, jc
  └───── one pair

Implemented for one pair and high card hands.